Updated player with animations and now is a mario.
Player is now a hand-drawn mario with resting, running/walking, and jumping animations. The script is updated to activate them in the correct ways. The score font color on the game over menu was updated since it was hard to see. Finally, when something hits a platform, it is checked to make sure that it is the player before adding a point. Photoshop importer package is now needed to import the different layers of the mario sprite.